chore: move pkg/sypgp to internal/pkg/sypgp
The sypgp package isn't really suitable for external consumption.

Key handling / management, including sygpg, is intended to go through
some refactoring in the near to mid future.

Move to internal at this point, while we are at a major version bump.

Part of #80
